requirements.txt
fastapi
uvicorn
python-dotenv

langchain
langchain-core
langchain-community
langchain-huggingface
langchain-groq
langchain-astradb

astrapy
cassio
faiss-cpu
pypdf
bs4
wikipedia